refactor: remove trendingContext and replace with useTrending and useGoToDetails hooks
