+++ src/Components/Hooks/useGoToDetails.js
@@ -1,0 +1,14 @@
+import { useNavigate } from "react-router-dom";
+
+export function useGoToDetails() {
+  const navigate = useNavigate();
+
+  function goToDetails(id, mediaType) {
+    navigate({
+      pathname: "/Details",
+      search: `?id=${id}&mediaType=${mediaType}`,
+    });
+  }
+
+  return { goToDetails };
+}
